`CNDIT-1103`: Move the patient and provider model objects to respective packages in the dto Fix the stream processing logic to surface the raw json data along with the transformed values in the final topic
