+++ data-reporting-service/person-service/src/main/java/gov/cdc/etldatapipeline/person/model/dto/provider/Provider.java
@@ -1,4 +1,4 @@
-package gov.cdc.etldatapipeline.person.model.dto;
+package gov.cdc.etldatapipeline.person.model.dto.provider;
 
 import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
 import com.fasterxml.jackson.core.JsonProcessingException;
